Support date separators in virtual message list
Adds disableDateSeparator property for the
virtual message list. To avoid breaking changes,
the default value of the property is true.
